extras: Use char instead of int for FileDescriptorActivity::CMD_xxx values.
Only the first byte of the FileDescriptorActivity::CMD_xxx values are
used by the software, so these values should only be a single byte in
length.  Defining them as 4-byte ints works on little-endian machines
because the first byte is the least-significant one.  On big-endian
machines, though, the first byte is the most-significant one, which
is zero for all three CMD_xx values.  Hence, all three CMD_xxx values
are interpreted as CMD_BREAK_LOOP on big-endian machines.