useorocos: collect compiler and linker flags in orocos_use_package
The orocos_use_package(...) function has been converted to a macro,
which additionally modifies the USE_OROCOS_COMPILE_FLAGS and
USE_OROCOS_LINKER_FLAGS to allow packages to export arbitrary
compiler and linker flags to be applied to each dependent orocos
target. Output variables are added to the internal cache.

Signed-off-by: Johannes Meyer <meyer@fsr.tu-darmstadt.de>