mqueue: factor the createStream in a base class
The issue this commit tries to fix is that all subclasses of
MQTemplateProtocol have the restriction that T needs to have
no virtual tables, as a static assert in MQTemplateProtocol
checks this.

However, there is no need for such a check in the general case,
only the "use memcpy to marshal" case has this issue. So,
create MQTemplateProtocolBase to get the completely generic
createStream method, and make the other marshaller implementations
depend on it