corba: give a way to forward C++ exceptions across the CORBA layer
Right now, every error in the RTT-CORBA convertions and/or access to the RTT
methods has to be translated into either a return value (if the method has
true/false) or a specific exception. This gets difficult when the problem comes
from sanity checks in the transport code itself.

This commit introduces a specific CORBA exception (StdException) which embeds
the C++ exception message (from what()), at least allowing to give proper
feedback to the caller.