Revert "services: also clear DataFlowInterface in Service::clear()"
This reverts commit 2a254afb04ef4c922d4a9bd80b15c05e9628762c.

The DataFlowInterface cannot be cleared anymore at this point as the port objects might already be destroyed.
`DataFlowInterface::removePort()` is purely based on names, but `DataFlowInterface::clear()` calls `getName()`
on the destroyed port object (invalid read).

For a clean solution we would have to cache the port names or the pointers to the created port objects in the
DataFlowInterface...

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>