extras: Support an intended period in FileDescriptorActivity
A component querying a FileDescriptorActivity (FDA) will always
receive back a period=0. This may be invalid if the FDA is actually
tracking a periodic signal, e.g. say a sync signal from hardware. In
this case there is a period at deployment time that the component
is intended to run at, but this is not available to the underlying
component.

This adds an intended period and returns that as the period of this
activity. The default intended period is the same as the actual
period=0. The actual period remains the same; 0.