service: fixup shared_from_this saga for setOwner/setParent.
We now catch the bad weak ptr and create the shared_ptr
ourselves. This is allowed since we know that it is stored
in the other Service and not destructed at the end of the
function.

In addition, for Boost 1.53+. use enable_shared_from_raw which
has the old enable_shared_from_this2 behavior again.

See http://stackoverflow.com/questions/22184772/why-is-boostenable-shared-from-raw-so-undocumented

Signed-off-by: Peter Soetens <peter@thesourceworks.com>