operations: introduce addEventOperation() and enable ORO_SIGNALLING_OPERATIONS by default.
The code/compilation time increase is purely because of the boost::fusion code,
so we eliminate generating that code in addOperation and enable it in addEventOperation.

In this way, we can provide binary compatible signalling operations without putting
the burden on users/components that don't use that feature.

This does introduce an extra shared_ptr in the operation base classes, so excluding
all this code can still be done by turning ORO_SIGNALLING_OPERATIONS off in cmake.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>