scripting: support responding to peer event operations
We lifted the limitation that signalling operations need to be ownthread.
The signal is now always executed in the thread of the reacting task,
The scripting StateMachine code uses this to support Orocos 1.x style
events. Each signalled operation call is dispatched to the reacting
task, together with the arguments of the operation. The limit of
possibly dispatched signals is the amount of available TLSF memory.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>