engine: make updateHook() and processFunctions() behave correctly again
This patch addresses issue #70 by a minor modification of TaskCore and
the ExecutionEngine. The big scheme throughout this patch are these 3
rules:

1. one uses task.setPeriod( s ) to set the period adn task.getPeriod() to
read the period

2. event port callbacks no longer schedule an updateHook(). Also completed
operation calls no longer schedule an updateHook.

3. one uses task.trigger() to schedule the execution of updateHook() or
processFunctions() in non-periodic tasks. This can be called from anywhere,
so from event port callbacks or external components.

4. one uses task.update() to directly execute an updateHook() in case of
a SlaveActivity.

Side effects:

A. For periodic components, event port callbacks will be called directly
in the thread of the component, no longer waiting for a period to expire.
This effectively decouples the reception of data on ports from updateHook(),
while still being fully thread-safe.

B. If rule 1. is followed, a task.getPeriod() will return 's', while a
task.engine()->getActivity()->getPeriod() will return '0.0'. This could be
fixed though...

C. For non-periodic components, the execution of an event port callback
will no longer schedule an updateHook(), unless task.trigger() is called
in that callback

D. For non-periodic components, script functions and updateHook() are
only called as many times as trigger() is called, with the semantics
that trigger() calls do not accumulate, but schedule at most one execution
of updateHook/processFunctions.

This patch only required a single modification in the unit tests: a
corba server had to schedule an updateHook() (effect C.) from an event port callback.

If your code does not follow convention '1.' above, the code will behave
as before for periodic components/activities (side effect A+B not present).
For non-periodic components, side effect C and D is always present.

This patch is a work in progress, and some minor follow-up patches may
follow.

Technical solution:

- We no longer let the Activity schedule a run of the ExecutionEngine,
but use the EE's condition variable to schedule periodic/non-periodic
task execution. Because of this, we can distinguish between a 'user'
task.trigger() and a 'thread/internal' ( = getActivity()->trigger()) trigger.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>