rtt: moved isRunning() check from TaskContext::dataOnPort() to default TaskContext::dataOnPortHook()
This allows the user to decide in an overwritten dataOnPortHook() implementation whether the user callbacks
should be enqueued or not, indepedent of whether the component is running or not. Note that the callbacks are
not executed until the component is started unless #70 is merged.

See https://github.com/orocos-toolchain/rtt/commit/d03c66b69912714b851818d471beb3282f19bad3#commitcomment-8247917.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>