scripting: fix .send() bug when using a SendHandle for exported script functions
We noticed that a send() on a script function did not work when
the return value was stored in a SendHandle. The reason was that the
SendHandleAlias was stored as an Attribute, but can't be instantiated
(since its an alias to a FusedFunctorDataSource), hence, it copied
all arguments, prior to giving the arguments a chance to instantiate.
so this send() then used copies that weren't used later on...
We now remove the SendHandleAlias from the StateMachineService such
that they are removed before instantiation happens. We only need them
during parsing as attributes, not during execution.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>