engine: run updateHook() only upon trigger or timeout
This patch is supposed to fix #70 in a similar, but improved way
as the original proposal.

The main introduction is the addition of RunnableInterface::work( WorkReason )
and ActivityInterface::timeout() . This allows us to inform the runnable
interface of the reason why it is being executed: because of a TimeOut,
a Trigger, or IO ready. Depending on the reason, the RunnableInterface
subclass can choose to do something different. The ExecutionEngine uses this
to distinguish between message bookkeeping (Triggers) and real work to do
in updateHook ( TimeOut and IOReady ).

In addition, when a TaskContext is Running, an EventPort will call updateHook()
upon newdata or, if a callback is attached, that callback and not updateHook.
This works for periodic and non-periodic *Activity implementations.

Unfortunately, the TaskContext::trigger() will forward to ActivityInterface::timeout()
and the ExecutionEngine will use ActivityInterface::trigger() internally to
do the message/operation bookkeeping. This is confusing in the implementation
side, but keeps the API fully backwards portable.

The implementation implements this mechanism for each Activity kind in RTT.

The side effects are minimal to none. We do call always loop() now from Thread,
and this patch does not yet detect Thread overruns (to be implemented in Activity).

Signed-off-by: Peter Soetens <peter@thesourceworks.com>