activity: fix pure virtual function calls at thread termination after destruction of Activity
This patch clears the Thread active flag in Thread::terminate() and forces the thread
function to be terminated in the destructor of class Activity. This solves some issues with pure virtual
method calls during the destruction phase if threads are still running and triggering each other.

It used to be part of a bigger commit 58cb4bb733fc14de0ca520620cfcb689e2155797, that has been reverted.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>