operations: improve OperationCallerInterface::isSend() for master-slave calls
This is a combination of 3 commits.

The first commit's message is:
operations: assert if the caller is not set and improve isSend

isSend now only returns true if the thread differs, not just the EE.
So that it also does not send in case of a master-slave call.

Signed-off-by: Peter Soetens <peter@intermodalics.eu>

This is the 2nd commit message:

operations: fixup isSend() logic to also work if no caller is known.

Signed-off-by: Peter Soetens <peter@intermodalics.eu>

This is the 3rd commit message:

operations: improved isSend implementation that does not need caller

As suggested by @meyerj, we can use isSelf() instead of comparing threads.

Signed-off-by: Peter Soetens <peter@intermodalics.eu>