extras: Add simulation of FileDescriptorActivity useful to unit tests
For unit tests that want to exercise a component which uses/requires a
FileDescriptorActivity (FDA), the simulation version of FDA can be used
to a) act as an actual FDA, and b) provide deterministic control of an
FDA and, in particular, when and how it wakes up the component under
test. See work(), isUpdated(), and hasTimeout(), in particular, for
behaviour related to deterministic control by a unit test.

Have to model both 'period' and 'running' to keep other parts of
RTT happy, in particular, the TaskContext/TaskCore implementation that,
in part, relies on isRunning() to make choices on what to do. The period
is mostly to keep the component under test happy, which may require that
it be running in a non-perioid activity.