Fixed construction and connection of shared connections with remote input ports
Allowing to connect remote input ports to an existing shared connections required to add a new API method createConnection(SharedConnectionBase::shared_ptr) to be added to the PortInterface class with a special
overwritten implementation in class RemoteInputPort. This call is only forwarded to the remote side, without building a new remote connection.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>