Added RTT v2 compatibility mode
If the RTT_V2_COMPATIBILITY_MODE or RTT_V2_DATAFLOW COMPATIBILITY_MODE macro is defined, the ConnFactory will install per-connection buffers again, even for push connections.
Additionally, an InputPort checks the policies of new connections and logs a meaningful error message if the connection attempt would have failed without the
compatibility mode.

For now this is a compile-time decision, but it would be possible to make it a run-time flag, too.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>