useorocos: check the real path of the current and project source dir for automatic Orocos package name derivation
This is required if at least one component of the path to the source directory is a symlink, because PROJECT_SOURCE_DIR
points to the symlinked location and CMAKE_CURRENT_SOURCE_DIR contains the real path of the project's source directory.

TODO: Abort with an error if CMAKE_CURRENT_SOURCE_DIR is not a subfolder of PROJECT_SOURCE_DIR and no package name was given
explicitly.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>