corba: return NotConnected after CORBA throwed an Exception during a write call
This is especially required if the other end of the connection disappeared, in order to properly cleanup
and to not spam the log with transient CORBA errors. If required, it is up to the user application to
reestablish the connection automatically in case the connection breaks.

Instead, we check the return value of updateAny/updateFromAny and return a WriteFailure in case conversion fails.
This is most likely a temporary error and not due to a broken CORBA connection.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>