engine: fixed regression from 7a8e48871e7b25301e942bbbd3cbdcc2cac35de7 that broke functions in slave engine
A previous patch "engine: fixed forwarding of messages and port callbacks received in slaves to the master engine"
(7a8e48871e7b25301e942bbbd3cbdcc2cac35de7) changed the RTT::ExecutionEngine()::waitForFunctions() method so that
it forwards the call to the master engine if set. While this is the intended behavior for operations/messages and
port callbacks, functions that finish execution signal the slave engine directly and hence waitForFunctions() must
not forward to the master, too.

The patch to waitForFunctions() is actually not required to fix message and port callback forwarding for slaves.
It was based on the wrong assumption that waitForMessagesInternal() must only be called if the ExecutionEngine has
no master.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>