slaveactivity: forward isPeriodic() member call to the master engine
In case the SlaveActivity has a master (mmaster != NULL), the slave's mperiod member variable
should not be used directly. We always return the periodicity of the master, especially because
it could have changed since the SlaveActivity was created and there is no API to either change
the master pointer or to set the slave's periodicity if a master was given in the constructor.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>