scripting: use message queue to call functions asynchronously (fix #150)
Instead of queueing a function in the function queue of the executing engine the new
implementation of RTT::scripting::CallFunction implements a DisposableInterface and
enqueues itself in the message queue which is processed asynchronously. This way the
engine does not have to wait for a timeout or trigger and execute as soon as possible.

The scripting_test is modified so that it would block indefinitely for the same
reason as for the example script in #150 (without SequentialActivity).

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>