corba: manage links to Service instances in weak pointers and copy task name for logging during cleanup
The CORBA layer should not delay the destruction of Service instances once the owning TaskContext gets destroyed.
Accessing a destroyed service instance through CORBA would result in a bad_weak_ptr exception.

For the case the TaskContext is destroyed before the TaskContextServer we need to make a copy of the task name
for logging in the destructor. There is no mechanism in RTT that guarantees that a server is destroyed before the
actual TaskContext.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>