property: remove copy() method and allow to construct or create a property with a custom data source
PropertyBag assignment is supposed to have linking semantics (changing the value of a property in the copy
also changes the value in the original bag), which is important for property composition and decomposition.

The PropertyBase::copy() method introduced in 44b9970e9685b9b812ddce1e8cc8c9b400553d54 had this linking
semantics and was used for PropertyBag assignment, but the term `copy` was actually always
associated with deep-copy semantics in RTT, like in PropertyBase::copy(const PropertyBase *) or in
DataSourceBase::copy().

This patch introduces an alternative signature for the create() method, which accepts a custom data source
instead of creating a default-initialized ValueDataSource. PropertyBag assignment now calls this create()
method with the DataSource pointer of the original property to be copied.

This reverts commit 44b9970e9685b9b812ddce1e8cc8c9b400553d54.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>