Fixed decomposition of PropertyBags which contain types that implement the TemplateCompositionFactory interface
RTT::types::decomposePropertyBag() did not retain the type of the PropertyBag returned by RTT::types::TypeInfo::decomposeType(),
which is implemented in subclasses of TemplateCompositionFactory, and therefore the subsequent composition did not recognize the
original type. The KDL typekit implements composition and decomposition using that method. The fallback method of property
decomposition implemented in types::propertyDecomposition() using getMember() is not affected by this issue and sets the type
of the target bag correctly.

The updated test in RTT cannot easily cover this case, because no RTT core typekit implements the TemplateCompositionFactory
interface.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>