Prevent activity changes from a component's own thread (fix #141)
This patch partially solves issue #141. It is not possible to change the activity of a component
from within the component's thread because it would have to destroy itself afterwards.
RTT::TaskContext::setActivity() logs an error and returns false in this case.

The debug log output during thread termination has been moved from the thread destructor to the
terminate() method to fix logging during the destruction of RTT::Activity (see also 58cb4bb).

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>