cmake: updated exact version check semantics for find_package(Orocos-RTT)
Every version that has the same major version number and at least the
requested minor version number is considered as compatible. Nothing changed here.

If an exact version of RTT is requested, the script also accepts it if the patch version
or the minor version was not specified. If they are, they have to match.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>