Revert "slaveactivity: forward isPeriodic() member call to the master engine"
The mmaster pointer might be dangling and accessing it from the isPeriodic()
call in SlaveActivity::stop() triggers a segfault if the master is already
destroyed.

Ideally, activities would be managed in shared pointers only and the slave
activity uses the weak pointer concept to check if the master is still available.

This reverts commit cb4a37687f2a8639381bfcd181cbb13ec0c01622.