fix race conditions between stop() and the runtime error state
stop() does a synchronization point by stopping the activity, but
then starts it right away. Given that the EE was not checking against
mTargetState when deciding to call errorHook(), there was a chance
that errorHook was called during or after stopHook.

Moreover, while error() was checking that mTargetState was a runtime
state, recover() was not, which could lead to a call to recover()
overriding mTargetState during the stop, leaving the task stopped
but in an inconsistent state.