logger: Ensure that the notification of real-time logging being enabled is actually logged
If you log and then turn on logging, then you will not get the logged item if RealTime
level has been selected. This is simply due to the check occurring before everything
has been enabled again.

The fix is to turn on logging and then log, ensuring that the logged item will
actually be logged (if the appropriate log level is already set).