internal: fixed conversion operator return type in internal::AStore<T> and internal::DataStore<T>
These types are not used in a const context and therefore const correctness is not an issue.
This patch partially reverts ccf678ca53fab373ad1fd65f137993b893517e41.

For consistency with the get() accessors the return type of internal::AStore<T> and internal::DataStore<T>
has been changed to T&. Unfortunately this change alone does not fix the compile error reported in
https://github.com/orocos-toolchain/rtt/pull/196.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>