cmake: set minimum required cmake version to 2.8
It was updated to 2.8.3 in 811759c2 and 6e3c8eec (for UseOROCOS-RTT.cmake), but
a test with Ubuntu Lucid and cmake 2.8.0 showed no apparent issues. We are not using
any feature that was introduced in 2.8.3 only. The main reason why the minimum was
set to 2.8.3 is that this is the version shipped with Ubuntu Precise and the minimum
version of most ROS packages and in the standard catkin template.

I also tested with cmake 2.6.0 and 2.6.2, but there were some issues (cmake bugs?)
and I think we can safely drop support for older versions.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>