Check availability of certain newer pthread and Boost features
pthread_setname_np and pthread_getname_np is not available on all supported
gnulinux platforms. We check availability in check_depend.cmake now and add
a preprocessor guard.

Boost Uuid was a new requirement introduced in #114, but Boost <1.42 is lacking
the Uuid component. Added a workaround using libuuid if installed or abort with
an error during the cmake run.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>