engine: immediately trigger an update step after a scripting function has been called and loop in waitAndProcessFunctions() (fix #150)
This would be another solution for the blocking function call issue reported in https://github.com/orocos-toolchain/rtt/issues/150.

Whenever a scripting function is called, we also trigger a full update cycle, as if TaskContext::trigger() has been called. This
is a no-op for periodic activities, but makes sure that the function gets executed soon otherwise without an explicit trigger.
Additionally, if the caller and the runner's thread are the same, functions will be processed in a loop now in waitAndProcessFunctions()
because there is nothing else we can do anyway to resolve the situation and otherwise the engine would have to rely on an external trigger.
The previous solutions which blocked on the msg_cond condition did not guarantee that functions execute at most once in an update cycle
either.

This patch also contains a small optimization for waitAndProcessMessages(), which checked the predicate twice within each cycle.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>