os: Change thread name compaction algorithm to produce more readable results
The algorithm currently takes only the last 15 characters of the
activity's name. With long component/activity names this can easily
result in the same name for multiple components. This is exacerbated
in CORBA dispatchers which append "Corba" on the end of the activity's
name, thus reducing the usable name size to 10 characters. So any
component names that end in "Configuration" will all have the same
thread name, "figurationCorba", rendering them unidentifable.

A compromise solution is to take the first 7 char's of the name, and
the last 7 char's of the name, and put the "~" char between them. Now
components like "RobotConfiguration" and "MyConfiguration" become
"RobotCon~uration" and "MyConfiguration" respectively, and their
CORBA dispatchers become "RobotCon~onCorba" and "MyConfig~onCorba"
respectively. These are more recognizable.

Added a test for <15 char names.
Made thread_name_len a constant.