corba: do not delete the dispatcher map mutex when the map is empty
As soon as we assume that these methods must be thread safe, we
cannot guarantee that no other thread will try a Get, Acquire or
Release while another is deleting the mutex.

We could possibly have a fairly complex refcounting scheme to allow
for deletion, but I would rather just leave the singleton mutex live.
If static C++ initialization rules were not so complicated, the mutex
would probably be a global variable anyways.