thread: do not overwrite the inherited CPU affinity by default
If no specific CPU affinity was given in the Thread or Activity constructor, the cpu_affinity
argument has a default value of 0. The gnulinux implementation of rtos_task_create()
forwarded this value to rtos_task_set_cpu_affinity() which resets the affinity attributes to
all CPUs in this case. On the other hand, if cpu_affinity set by the user is ~0, the call to
rtos_task_set_cpu_affinity() was skipped and the new threads inherits the affinity of its
creator, the default behavior of pthread_create().

This patch basically swaps these two cases, so that new threads inherit the affinity of their
parent (typically the main thread) and the affinity is only overwritten in case the user sets
it explicitly. If the cpu_affinity argument is ~0, the affinity will be reset to all available
CPUs, disregarding the flags of the parent.

This might even be considered as a bug fix.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>