cmake: unset found and missing components list after each invocation of the RTT cmake config
If the RTT cmake module is used more than once with different components and the REQUIRED argument was not
given during the first invocation, a second invocation with REQUIRED might trigger false errors because of
items added to MISSING_COMPONENTS earlier.

Also moved initialization of OROCOS-RTT_PLUGIN_PATH to the if(NOT OROCOS-RTT_FOUND) section to not readd
the same directories over and over again.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>