corba: fixed unused-but-set-parameter warning in AnyConversion<std::string>::toCorbaType()
This non-templated overload of the method is not different from the template method, so can be
removed. Note that type `dummy` in the templated variant always needs to have reference semantics,
which is the case when called from AnyConversion<std::vector<std::string>>::toCorbaType() in
CorbaConversion.hpp:237.

Fixed compiler warning:
RTTCorbaConversion.hpp:267:41: warning: parameter ‘dest’ set but not used [-Wunused-but-set-parameter]

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>