tests: fixed execution thread spec in fixture for the state_test
The operation "vo0" added for the testStateYieldbyCmd check in state_test.cpp was not
an OwnThread operation and hence executed by the GlobalEngine instead of the task that
is also running the state machine. This patch does not change the result of the test:
.cmd() correctly yields and allows the engine to execute the operation before checking
the result in the next cycle.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>