tests: removed invalid TaskContext inheritance in corba_ipc_test.cpp and unused header corba_test.hpp
Inheritance of the fixture from TaskContext is invalid because it does
not cause the test body to be executed in the thread of the TaskContext's
activity. As a consequence, the specified caller this->engine() for the
OperationCallers in callBackPeer() was and still is incorrect, except for the
case when it is executed as the implemention of the callBackPeerOwn operation
owned by the fixture TaskContext.

The fact that the fixture inherited from TaskContext caused compile errors with
Boost 1.65.1 because the unit test framework tests for existence of a member called
setup, which is defined as private in RTT::TaskContext::setup().

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>