scripting: fix unexpected behavior of yield statements within a function
A yield statement within a function does not have a well-defined meaning if the function was called
from the same thread that is supposed to execute it, especially if it is nested within
another function. The ExecutionEngine::waitAndProcessFunctions() method, that was called in this
case, processed all functions running in the same engine again, even if not related to the ongoing call,
but did not process operations or invoke updateHook() before it returns. This is totally unexpected
behavior because

- it's different from what a yield statement would do in programs, functions being sent or if the function
  is called from a different thread
- other waiting yields are unblocked and state machines are stepped again within the same update cycle
  because of the intermingled calls to ExecutionEngine::processFunctions().

With this patch a yield statement becomes almost a no-op in this case and only processes the message queue
before it returns, which handles pending operation calls. Other than before the patch in
32bdc3c8668456c533e3f3d6c49e1766fb2c55cf, that made all function calls execute immediately without the need
to wait for the next update cycle, it is not required anymore to process *all* functions again while waiting
for the completion of the current one. ExecutionEngine::processFunctions() is called exactly once per update
step.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>